+++ Lib/TSCompiler/BundleBundler.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.IO;
 using Lib.CSSProcessor;
 using Lib.DiskCache;
+using System.Globalization;
 
 namespace Lib.TSCompiler
 {
@@ -13,6 +14,7 @@ namespace Lib.TSCompiler
     {
         string _mainJsBundleUrl;
         string _bundlePng;
+        List<float> _bundlePngInfo;
         string _indexHtml;
         readonly IToolsDir _tools;
 
@@ -72,7 +74,12 @@ namespace Lib.TSCompiler
                 var bundlePngContent = Project.SpriteGenerator.BuildImage(true);
                 if (bundlePngContent != null)
                 {
-                    FilesContent[_bundlePng] = bundlePngContent;
+                    _bundlePngInfo = new List<float>();
+                    foreach (var slice in bundlePngContent)
+                    {
+                        FilesContent[PathUtils.InjectQuality(_bundlePng, slice.Quality)] = slice.Content;
+                        _bundlePngInfo.Add(slice.Quality);
+                    }
                 }
                 else
                 {
@@ -126,6 +133,17 @@ namespace Lib.TSCompiler
             if (_bundlePng != null)
             {
                 res += $"var bobrilBPath=\"{_bundlePng}\"";
+                if (_bundlePngInfo.Count > 1)
+                {
+                    res += $",bobrilBPath2=[";
+                    for (int i = 1; i < _bundlePngInfo.Count; i++)
+                    {
+                        var q = _bundlePngInfo[i];
+                        if (i > 1) res += ",";
+                        res += $"[\"{PathUtils.InjectQuality(_bundlePng, q)}\",{q.ToString(CultureInfo.InvariantCulture)}]";
+                    }
+                    res += "]";
+                }
             }
             res += "</script>";
             return res;