+++ Lib/TSCompiler/FastBundleBundler.cs
@@ -2,9 +2,8 @@
 using System.Text;
 using Lib.Utils;
 using Lib.ToolsDir;
-using System;
 using System.Linq;
-using System.IO;
+using System.Globalization;
 
 namespace Lib.TSCompiler
 {
@@ -28,7 +27,8 @@ namespace Lib.TSCompiler
 
         // value could be string or byte[] or Lazy<string|byte[]>
         public Dictionary<string, object> FilesContent;
-        private string _bundlePng;
+        string _bundlePng;
+        List<float> _bundlePngInfo;
 
         public void Build(string sourceRoot, string mapUrl, bool testProj = false)
         {
@@ -76,7 +76,12 @@ namespace Lib.TSCompiler
                 var bundlePngContent = Project.SpriteGenerator.BuildImage(false);
                 if (bundlePngContent != null)
                 {
-                    FilesContent[_bundlePng] = bundlePngContent;
+                    _bundlePngInfo = new List<float>();
+                    foreach (var slice in bundlePngContent)
+                    {
+                        FilesContent[PathUtils.InjectQuality(_bundlePng, slice.Quality)] = slice.Content;
+                        _bundlePngInfo.Add(slice.Quality);
+                    }
                 }
                 else
                 {
@@ -208,6 +213,17 @@ namespace Lib.TSCompiler
             if (_bundlePng != null)
             {
                 res += $"var bobrilBPath=\"{_bundlePng}\"";
+                if (_bundlePngInfo.Count > 1)
+                {
+                    res += $",bobrilBPath2=[";
+                    for (int i = 1; i < _bundlePngInfo.Count; i++)
+                    {
+                        var q = _bundlePngInfo[i];
+                        if (i > 1) res += ",";
+                        res += $"[\"{PathUtils.InjectQuality(_bundlePng, q)}\",{q.ToString(CultureInfo.InvariantCulture)}]";
+                    }
+                    res += "]";
+                }
             }
             res += "</script>";
             return res;